requirements.txt
# Core dependencies
numpy>=1.24.0
pyglet>=2.0.0

# GPU acceleration (macOS)
pyobjc-framework-Metal>=9.0; sys_platform == "darwin"
pyobjc-framework-MetalKit>=9.0; sys_platform == "darwin"

# API server
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Optional dependencies
scipy>=1.10.0
